splitting simplified_interface.h in two files to make it easier to read for users
